[353] 개인 임계치 수정 요청 필드 타입 수정

+++ src/main/java/com/pooli/line/mapper/LineMapper.java
@@ -33,7 +33,7 @@ public interface LineMapper {
 	
 	int updateIndividualThreshold(
 	          @Param("lineId") Long lineId,
-	          @Param("individualThreshold") Integer individualThreshold,
+	          @Param("individualThreshold") Long individualThreshold,
 	          @Param("isThresholdActive") Boolean isThresholdActive
 	  );
